% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{addSimilarities}
\alias{addSimilarities}
\title{Additive integration of similarities}
\usage{
addSimilarities(x, bio_mat, weights = c(0.5, 0.18, 0.1, 0.22))
}
\arguments{
\item{x}{A matrix with the similarity of expression}

\item{bio_mat}{A list of matrices of the same dimension}

\item{weights}{A numeric vector of weight to multiply each similarity}
}
\value{
A similarity matrix
}
\description{
Function that used the previously calculated similarities into a single
similarity matrix.
}
\details{
The total weight sum can't be higher than 1 to prevent values above
1 but can be below 1.
}

