% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_GeneSetCollection.R
\docType{methods}
\name{as.GeneSetCollection}
\alias{as.GeneSetCollection}
\alias{as.GeneSetCollection,list-method}
\alias{as.GeneSetCollection,AnnDbBimap-method}
\title{Convert a list or a Bimap interface into GeneSetCollections}
\usage{
as.GeneSetCollection(object)

\S4method{as.GeneSetCollection}{list}(object)

\S4method{as.GeneSetCollection}{AnnDbBimap}(object)
}
\arguments{
\item{object}{A list of pathways, genes or an AnnDbBimap.}
}
\value{
A GeneSetCollection
}
\description{
Convert a list or a Bimap interface into GeneSetCollections
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Convert list to GeneSetCollections

\item \code{AnnDbBimap}: Convert AnnDbBimap to GeneSetCollections
}}

\author{
Lluís Revilla
}
