% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{bio.cor2}
\alias{bio.cor2}
\title{Comparing information in databases}
\usage{
bio.cor2(genes_id, ids = "Entrez Gene", go = FALSE, react = TRUE,
  kegg = FALSE, all = FALSE)
}
\arguments{
\item{genes_id}{is vector of ids to compare}

\item{ids}{indicate if the id is eihter Entre Gene or Symbol}

\item{go}{logical; indicates if the overlap in terms of GO is calculated}

\item{react}{logical; indicates if the overlap in Reactome pathway is
calculated}

\item{kegg}{logical; indicates if the overlap in Kegg database is calculated}
}
\value{
A list where each element is a matrix with the similarity for such
database
}
\description{
Calculates a functional similarity of genes  using information available in
several databases.
}
\details{
For Gene Ontologies, the DAG path structure is used to compute how similar two
genes are. For metabolic pathways the max number of proteins involved in a
pathway for each gene is calculated.
}

