% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{bioCor}
\alias{bioCor}
\title{Comparing information in databases}
\usage{
bioCor(genes_id, ids = "Entrez Gene", react = TRUE, kegg = FALSE,
  all = FALSE)
}
\arguments{
\item{genes_id}{is vector of ids to compare}

\item{ids}{indicate if the id is eihter Entre Gene or Symbol}

\item{react}{logical; indicates if the similarities in Reactome pathway is
calculated}

\item{kegg}{logical; indicates if the similarities in Kegg database
is calculated}

\item{all}{logical; indicates if all the previous (go, react, and kegg)
similarity measures, should be set to TRUE.}
}
\value{
A list where each element is a matrix with the similarity for such
database \code{NA} indicates that there isn't any information of one of
those genes.
}
\description{
Calculates a functional similarity of genes  using information available in
several databases.
}
\details{
For Gene Ontologies, the DAG path structure is used to compute how similar
two genes are. For metabolic pathways the max number of proteins involved in
a pathway for each gene is calculated.
}

