% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{comb_biopath}
\alias{comb_biopath}
\title{Extract all the ids of biopath for each element on the combination}
\usage{
comb_biopath(comb, info, by, biopath)
}
\arguments{
\item{comb}{ids of the column "by" to compare}

\item{info}{is the data.fram with information of the by and biopath columns}

\item{by}{Type of columns to subset}

\item{biopath}{Column of info we want to compare}
}
\value{
A matrix of combinations of all the ids selected from biopath
and compare them all
}
\description{
Extract all the ids of biopath for each element on the combination
}

