% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{combBiopath}
\alias{combBiopath}
\title{Wrapper to expand.grid for pathways of genes}
\usage{
combBiopath(comb, info, by, biopath)
}
\arguments{
\item{comb}{Pair of ids of genes to find the pathways of. They should be on
the column indicated by \code{by} to compare}

\item{info}{The data.fram with a column for the genes id, and another for
pathway}

\item{by}{Type of columns to subset}

\item{biopath}{Column of info we want to compare}
}
\value{
A matrix of combinations of all the ids selected from biopath
and compare them all
}
\description{
Given a data.frame with the information and the genes we want to extract,
finds all the pathway for each gene and return a data.frame with the
combinations of all pathways of gene 1 with all pathways of gene 2
}
\author{
Lluís Revilla
}
