% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{conversions}
\alias{conversions}
\alias{D2J}
\alias{J2D}
\title{Convert the similarities formats}
\usage{
D2J(D)

J2D(J)
}
\arguments{
\item{D}{Dice coefficient, as returned by \code{\link{diceSim}},
\code{\link{geneSim}}, \code{\link{clusterSim}}, \code{\link{clusterGeneSim}}
 and \code{\link{bioCor}}}

\item{J}{Jaccard coefficient}
}
\value{
A numeric value.
}
\description{
Functions to convert the similarity coefficients between Jaccard and Dice.
D2J is the opposite of J2D.
}
\examples{
D2J(0.5)
J2D(0.5)
D2J(J2D(0.5))
}
\author{
Lluís Revilla
}
