% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{corPathways}
\alias{corPathways}
\title{Calculates the similarity score of patwhays}
\usage{
corPathways(comb, genes, id, pathwayDB)
}
\arguments{
\item{comb}{are the ids to compare, it is expected two ids of genes.}

\item{genes}{is the matrix with the information about "id" and "react"}

\item{id}{is the column of "genes" where \code{comb} are to be found}

\item{pathwayDB}{is the column of \code{genes} where pathways should be
found. It is usually the name of the database where they come from.}
}
\value{
The highest score of all the combinations of pathways between the
two ids compared.
}
\description{
Given the information about the relationship between the genes and the
pathways, uses the ids of the genes in comb to find the similarity score in
them.
}
\details{
If an \code{NA} is returned this means that there isn't information
available for any pathways for one of those two genes. Otherwise a number
between 0 and 1 (both included) is returned. Note that there isn't a negative
value of similarity for pathways correlation.
}

