% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-incidence.R
\docType{methods}
\name{distribution}
\alias{distribution}
\alias{distribution,GeneSetCollection-method}
\title{Distribution of genes per pathway in pathways}
\usage{
distribution(object)

\S4method{distribution}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection.}
}
\value{
A matrix with the number of pathways per genes as columns and the
genes per pathways as rows.
}
\description{
Calculates how many pathways per gene are distributed in genes per pathways.
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Returns the distribution of genes by size of the
pathways
}}

\seealso{
\code{\link{GenesPerPathway}}, \code{\link{PathwaysPerGene}}
}
\author{
Lluís Revilla
}
