% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{duplicateIndices}
\alias{duplicateIndices}
\title{Finds the indices of the duplicated events of a vector}
\usage{
duplicateIndices(vec)
}
\arguments{
\item{vec}{Vector of identififiers presumably duplicated}
}
\value{
The format is determined by the simplify2array
}
\description{
Finds the indices of duplicated elements in the vector given.
}
\details{
For each duplication it can return a list or if all the duplication events
are of the same length it returns a matrix, where each column is duplicated.
}
\examples{
duplicateIndices(c("52", "52", "53", "55")) # One repeated element
duplicateIndices(c("52", "52", "53", "55", "55")) # Repeated elements
duplicateIndices(c("52", "55", "53", "55", "52")) # Mixed repeated elements
}
\seealso{
\code{\link{removeDup}}
}
\author{
Lluís Revilla
}
