% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-methods.R
\docType{methods}
\name{gene}
\alias{gene}
\alias{gene,GeneSetCollection,character-method}
\alias{gene,GeneSetCollection,missing-method}
\title{Statistics about genes}
\usage{
gene(object, gene)

\S4method{gene}{GeneSetCollection,character}(object, gene)

\S4method{gene}{GeneSetCollection,missing}(object, gene)
}
\arguments{
\item{object}{A GeneSetCollection object}

\item{gene}{A character of the gene to analyze. If missing it will analyze
all genes}
}
\value{
If a single gene is provided invisible a list of statistics (the
statistics will be printed on the screen), otherwise the data.frame of
statistics.
}
\description{
Calculates some statistics of the genes, like the biggest
pathway, the gene in more pathways, h-index and entropy or information
content for the genes per pathway and pathways for genes and they percentage
of the maximum.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,gene = character}: Calculates statistics for a single gene

\item \code{object = GeneSetCollection,gene = missing}: Calculates statistics for all genes
}}

\author{
Lluís Revilla
}
