% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{gene2Sim}
\alias{gene2Sim}
\title{Add a similarity measure for a gene to existing similarities}
\usage{
gene2Sim(gene, func, sim, ...)
}
\arguments{
\item{gene}{Entrez Identifier of the gene to add the similarity calculation}

\item{func}{Function to calculate the similarity}

\item{sim}{Matrix with similarities already calculated.}

\item{...}{Other arguments passed to the function to calculate similarity}
}
\value{
A matrix with the similarities with the new similarity
}
\description{
Given a gene and a function to calculate the similarity with add it to the
previously calculated similarities in sim.
}
