% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{genes.info}
\alias{genes.info}
\title{Extract which genes are from which reactome}
\usage{
genes.info(genes, colm, id, type)
}
\arguments{
\item{genes}{is the data.frame with information}

\item{colm}{is the colum where "id" is found}

\item{id}{is the ids we are looking for in column Symbol of such data.frame}

\item{type}{is the column we are looking to keep.}
}
\value{
A vector with the unique identifiers of genes of the \code{type}
column
}
\description{
Extract which genes are from which reactome
}

