% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{genesInfo}
\alias{genesInfo}
\title{Extract which genes are from which database}
\usage{
genesInfo(genes, colm, id, type)
}
\arguments{
\item{genes}{is the data.frame with information}

\item{colm}{is the colum where \code{id} is found}

\item{id}{is the ids we are looking for in column \code{type} of
\code{genes}  data.frame}

\item{type}{is the column we are looking to keep.}
}
\value{
A list with the unique identifiers of genes of the \code{type}
column
}
\description{
Extract which genes are from which database
}
\examples{
library("org.Hs.eg.db")
library("reactome.db")
entrezids <- keys(org.Hs.eg.db, keytype = "ENTREZID")
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- select(org.Hs.eg.db, keys = entrezids, keytype = "ENTREZID",
                     columns = "PATH")
g <- genesInfo(genes.kegg, "PATH", c("04510", "04520"), "ENTREZID")
lengths(g)
}
