% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\name{GenesPerPathway}
\alias{GenesPerPathway}
\title{Genes per pathway}
\usage{
GenesPerPathway(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
a list of genes per pathway
}
\description{
Calculates the genes per pathway of a GeneSetCollection
}
\author{
Lluís Revilla
}
