% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\name{GenesPerPathway}
\alias{GenesPerPathway}
\title{Genes per pathway}
\usage{
GenesPerPathway(object)
}
\arguments{
\item{incidence}{Incidence matrix}
}
\value{
The number of genes each pathway has
}
\description{
Calculate the genes per pathway of an incidence matrix
}
\keyword{internal}
