% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{genesSim}
\alias{genesSim}
\title{Calculates the similarity score of two genes}
\usage{
genesSim(gene1, gene2, genes, id, pathwayDB)
}
\arguments{
\item{gene1}{Entrez gene id.}

\item{gene2}{Entrez gene id.}

\item{genes}{is the matrix with the information to calculate the similarity.
It is created using select, should contain the column "id" and "pathwayDB".}

\item{id}{is the column of "genes" where \code{gene1} and \code{gene2} are
to be found}

\item{pathwayDB}{is the column where pathways should be found. It is usually
the name of the database where they come from.}
}
\value{
The highest score of all the combinations of pathways between the
two ids compared.
}
\description{
Given the information about the relationship between the genes and the
pathways, uses the ids of the genes in comb to find the similarity score in
them.
}
\details{
If an \code{NA} is returned this means that there isn't information
available for any pathways for one of those two genes. Otherwise a number
between 0 and 1 (both included) is returned. Note that there isn't a
negative value of similarity for pathways correlation.
}
\examples{
library("org.Hs.eg.db")
library("reactome.db")
entrezids <- keys(org.Hs.eg.db, keytype = "ENTREZID")
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- select(org.Hs.eg.db, keys = entrezids, keytype = "ENTREZID",
                     columns = "PATH")
genesSim("81", "18", genes.kegg, "ENTREZID", "PATH")
# Extracts the paths of all genes of org.Hs.eg.db from reactome
genes.react <- select(reactome.db, keys = entrezids, keytype = "ENTREZID",
                      columns = "REACTOMEID")
genesSim("81", "18", genes.react, "ENTREZID", "REACTOMEID")
}
\author{
Lluis Revilla
}
