% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{goCor}
\alias{goCor}
\title{goCor}
\usage{
goCor(genes, semData, measure = "Wang", drop = "IEA", combine = "BMA",
  verbose = TRUE)
}
\description{
Functions that mimicks mgeneSim of GOSemSim package, but without filtering
NAs
}

