% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{h_index}
\alias{h_index}
\title{Calculates H-index}
\usage{
h_index(x)
}
\arguments{
\item{x}{The table or values to calculate the h-index from it.}
}
\value{
The numeric value of the h.index
}
\description{
Given a table find which is the h-index of the data
}
\keyword{internal}
