% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{inverseList}
\alias{inverseList}
\title{Invert a list}
\usage{
inverseList(x)
}
\arguments{
\item{x}{A list with genes as names and names of pathways as values of the
list}
}
\value{
The number of pathways each gene has.
}
\description{
Calculate the pathways per gene of list
}
\author{
Lluís Revilla
}
\keyword{internal}
