% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{mgene2Sim}
\alias{mgene2Sim}
\title{Add a similarity measure for a list of genes to existing similarities}
\usage{
mgene2Sim(gene, func, sim, ...)
}
\arguments{
\item{gene}{A list of Entrez Identifier of the gene to add to the similarity
calculation}

\item{func}{Function to calculate the similarity}

\item{sim}{Matrix with similarities already calculated.}

\item{...}{Other arguments passed to the functio to calculate the
similarities}
}
\value{
A matrix with the similarities with the new similarity
}
\description{
Given a gene and a function to calculate the similarity with add it to the
previously calculated similarities in sim.
}

