% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway-method.R
\docType{methods}
\name{pathway}
\alias{pathway}
\alias{pathway,GeneSetCollection,character-method}
\alias{pathway,GeneSetCollection,missing-method}
\title{Statistics about pathways}
\usage{
pathway(object, pathway)

\S4method{pathway}{GeneSetCollection,character}(object, pathway)

\S4method{pathway}{GeneSetCollection,missing}(object, pathway)
}
\arguments{
\item{object}{A GeneSetCollection object}

\item{pathway}{A character of the pathway to analyze. If missing it will analyze
all pathways}
}
\value{
If a single pathway is provided invisible a list of statistics (the
statistics will be printed on the screen), otherwise the data.frame of
statistics.
}
\description{
Calculates some statistics of the genes, like the biggest
pathway, the gene in more pathways, h-index and entropy or information
content for the genes per pathway and pathways for genes and they percentage
of the maximum.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = GeneSetCollection,pathway = character}: Calculates statistics for a single pathway

\item \code{object = GeneSetCollection,pathway = missing}: Calculates statistics for all pathways
}}

\author{
Lluís Revilla
}
