% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\name{PathwaysPerGene}
\alias{PathwaysPerGene}
\title{Pathways per gene}
\usage{
PathwaysPerGene(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
a list of pathways per genes
}
\description{
Calculates the pathways per gene of a GeneSetCollection
}
\author{
Lluís Revilla
}
