% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\name{PathwaysPerGene}
\alias{PathwaysPerGene}
\title{Pathways per gene}
\usage{
PathwaysPerGene(object)
}
\arguments{
\item{incidence}{Incidence matrix}
}
\value{
The number of pathways each gene has
}
\description{
Calculate the pathways per gene of an incidence matrix
}
\keyword{internal}
