% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{react_genes}
\alias{react_genes}
\title{Calculates the similarity score of patwhays}
\usage{
react_genes(comb, genes, react, id)
}
\arguments{
\item{comb}{are the ids to compare}

\item{genes}{is the matrix with the information about "id" and "react"}

\item{react}{is the column of \code{genes} where pathways should be found.}

\item{id}{is the column of "genes" where comb are to be found}
}
\value{
the max score of comparing the ids of such reaction
}
\description{
Given the information about the relationship between the genes and the
pathways, uses the ids of the genes in comb to find the similarity score in
them.
}

