% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{removeDup}
\alias{removeDup}
\title{Remove duplicated rows and columns}
\usage{
removeDup(cor_mat, dupli)
}
\arguments{
\item{dupli}{List of indicies with duplicated entries}

\item{mat}{List of matrices}
}
\value{
A matrix with only one of the columns and rows duplicated
}
\description{
Given the indices of the duplicated entries remove the columns and rows until
 just one is left, it keeps the duplicated with the highest absolute mean
value.
}

