% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{rm.dup}
\alias{rm.dup}
\title{Remove duplicates based on the higher similarity}
\usage{
rm.dup(adj, bio_mat, adj.key = "SYMBOL", bio_mat.key = NULL)
}
\arguments{
\item{adj}{Adjacency matrix}

\item{bio_mat}{List of matrices with similarities. Usually from functional
similarities}

\item{adj.key}{Original identifier used in the adjacency matrix.}

\item{bio_mat.key}{Identifier used in the similarities.}
}
\value{
A list of similarities as \code{bio_mat} but each from the same dimension as
\code{adj}.
}
\description{
When mapping Symbols to Entrez Identifiers there isn't a 1 to 1
correspondence. In this case to convert back to the user provided
identifiers this function can be used. Select the genes with the highest
similarities between the identifiers mapped to the same original identifier
}

