% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{setCores}
\alias{setCores}
\title{Parallel power}
\usage{
setCores(n = 2)
}
\arguments{
\item{n}{Number of cores used}
}
\value{
The result of registerDoParallel
}
\description{
A wrapper to registerDoParallel code
}
\seealso{
\code{\link[doParallel]{registerDoParallel}}
}

