% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-method.R
\docType{methods}
\name{summary,GeneSetCollection-method}
\alias{summary,GeneSetCollection-method}
\title{Calculates some statistics of the GeneSetCollection}
\usage{
\S4method{summary}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
invisible a list of statistics
}
\description{
These statistics include the biggest pathway, the gene in more pathways,
h-index and entropy or information content for the genes per pathway and
pathways for genes and they percentage  of the maximum.
}
\seealso{
\code{\link{GenesPerPathway}}, \code{\link{PathwaysPerGene}}
}
\author{
Lluís Revilla
}
