% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{weighted.sum}
\alias{weighted}
\alias{weighted.prod}
\alias{weighted.sum}
\title{Calculates the weighted}
\usage{
weighted.sum(x, w)

weighted.prod(x, w)
}
\arguments{
\item{x}{an object containing the values whose weighted operations is to be
computed}

\item{w}{a numerical vector of weights the same length as \code{x} giving
    the weights to use for elements of \code{x}.}
}
\value{
\code{weighted.sum} returns the sum of the product of x*weights
removing all \code{NA} values

\code{weighted.prod} returns the product of product of x*weights removing all \code{NA} values
}
\description{
Calculates the weighted sum or product of \code{x}. Each values should have
its weight, otherwise it will throw an error.
}

