% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_GeneSetCollection.R
\name{as.GeneSetCollection}
\alias{as.GeneSetCollection}
\title{Convert a list or a Bimap interface into GeneSetCollections}
\usage{
as.GeneSetCollection(object)
}
\arguments{
\item{object}{A list of pathways, genes or an AnnDbBimap.}
}
\value{
A GeneSetCollection
}
\description{
Convert a list or a Bimap interface into GeneSetCollections
}
\author{
Lluís Revilla
}
