% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioCor.R
\name{bioCor}
\alias{bioCor}
\title{Comparing information in databases}
\usage{
bioCor(genes_id, ids = "ENTREZID", react = TRUE, kegg = FALSE,
  all = FALSE, BPPARAM = bpparam())
}
\arguments{
\item{genes_id}{is vector of ids to compare}

\item{ids}{indicate if the id is eihter ENTREZID or SYMBOL}

\item{react}{logical; indicates if the similarities in Reactome pathway is
calculated}

\item{kegg}{logical; indicates if the similarities in Kegg database
is calculated}

\item{all}{logical; indicates if all the previous (go, react, and kegg)
similarity measures, should be set to TRUE.}

\item{BPPARAM}{
    An optional \code{\link{BiocParallelParam}} instance
    determining the parallel back-end to be used during evaluation, or a
    \code{list} of \code{BiocParallelParam} instances, to be applied in
    sequence for nested calls to \pkg{BiocParallel} functions.
  }
}
\value{
A list where each element is a matrix with the similarity for such
database \code{NA} indicates that there isn't any information of one of
those genes for that pathway database. 0 If there is information of both
genes but no overlap between those genes.
}
\description{
Calculates a functional similarity of genes  using information available in
several databases.
}
\examples{
bioCor(c("18", "81"))
}
\seealso{
\code{\link{geneSim}}
}
\author{
Lluís Revilla
}
