% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGeneSim.R
\name{clusterGeneSim}
\alias{clusterGeneSim}
\alias{mclusterGeneSim}
\title{Similarity score between clusters of genes based on genes similarity}
\usage{
clusterGeneSim(cluster1, cluster2, info, method = c("max", "rcmax.avg"), ...)

mclusterGeneSim(clusters, info, method = c("max", "rcmax.avg"), ...)
}
\arguments{
\item{cluster1}{A vector with genes.}

\item{cluster2}{A vector with genes.}

\item{info}{A list of genes and the pathways they are involved.}

\item{method}{A vector with two  or one argument to be passed to
combineScores the first one is used to summarize the similarities of genes,
the second one for clusters.}

\item{...}{Other arguments passed to \code{\link{combineScoresPar}}}

\item{clusters}{A list of clusters of genes to be found in \code{id}.}
}
\value{
\code{clusterGeneSim} returns a similarity score of the two clusters
or the similarity between the genes of the two clusters.

\code{mclusterGeneSim} returns a matrix with the similarity scores
for each cluster comparison.
}
\description{
Looks for the similarity between genes of a group and then between each
group.
}
\details{
Differs with clusterGeneSim that first each combination between genes is
calculated, and with this values then the comparison between the two
clusters is done. Thus applying combineScores twice, one at gene level and
another one at cluster level.
}
\examples{
if (require("org.Hs.eg.db")) {
    #Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
    # data of June 31st 2011)
    genes.kegg <- as.list(org.Hs.egPATH)
    clusterGeneSim(c("18", "81", "10"), c("100", "10", "1"), genes.kegg)
    clusterGeneSim(c("18", "81", "10"), c("100", "10", "1"), genes.kegg,
                   c("avg", "avg"))
    clusterGeneSim(c("18", "81", "10"), c("100", "10", "1"), genes.kegg,
                   c("avg", "rcmax.avg"))
    (clus <- clusterGeneSim(c("18", "81", "10"), c("100", "10", "1"),
                            genes.kegg, "avg"))
    combineScores(clus, "rcmax.avg")
} else {
    warning('You need org.Hs.eg.db package for this example')
}

clusters <- list(cluster1 = c("18", "81", "10"),
                 cluster2 = c("100", "594", "836"),
                 cluster3 = c("18", "10", "83"))
mclusterGeneSim(clusters, genes.kegg)
mclusterGeneSim(clusters, genes.kegg, c("max", "avg"))
mclusterGeneSim(clusters, genes.kegg, c("max", "BMA"))
}
\seealso{
\code{\link{clusterGeneSim}}, \code{\link{combineScores}} and
\code{\link{conversions}}
}
\author{
Lluís Revilla
}
