% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSim.R
\name{clusterSim}
\alias{clusterSim}
\alias{mclusterSim}
\title{Similarity score between clusters of genes based on pathways similarity}
\usage{
clusterSim(cluster1, cluster2, info, method = "max", ...)

mclusterSim(clusters, info, method = "max", ...)
}
\arguments{
\item{cluster1, cluster2}{A vector with genes.}

\item{info}{A list of genes and the pathways they are involved.}

\item{method}{To combine the scores of each pathway, one of \code{c("avg",
"max", "rcmax", "rcmax.avg", "BMA")}, if NULL returns the matrix of
similarities.}

\item{...}{Other arguments passed to \code{\link{combineScores}}}

\item{clusters}{A list of clusters of genes to be found in \code{id}.}
}
\value{
\code{clusterSim} returns a similarity score of the two clusters

\code{mclusterSim} returns a matrix with the similarity scores for
each cluster comparison.
}
\description{
Looks for the similarity between genes in groups
}
\details{
Once the pathways for each cluster are found they are combined using
combineScores.
}
\examples{
if (require("org.Hs.eg.db")) {
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- as.list(org.Hs.egPATH)
clusterSim(c("9", "15", "10"), c("33", "19", "20"), genes.kegg)
clusterSim(c("9", "15", "10"), c("33", "19", "20"), genes.kegg, NULL)
clusterSim(c("9", "15", "10"), c("33", "19", "20"), genes.kegg, "avg")
} else {
warning('You need org.Hs.eg.db package for this example')
}

clusters <- list(cluster1 = c("18", "81", "10"),
                 cluster2 = c("100", "10", "1"),
                 cluster3 = c("18", "10", "83"))
mclusterSim(clusters, genes.kegg)
mclusterSim(clusters, genes.kegg, "avg")
}
\seealso{
For a different approach see \code{\link{clusterGeneSim}},
\code{\link{combineScores}} and \code{\link{conversions}}
}
\author{
Lluís Revilla
}
