% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{combBiopath}
\alias{combBiopath}
\title{Wrapper to expand.grid for pathways of genes}
\usage{
combBiopath(comb, info, by, biopath)
}
\arguments{
\item{comb}{Pair of ids of genes to find the pathways of. They should be on
the column indicated by \code{by} to compare}

\item{info}{The data.fram with a column for the genes id, and another for
pathway}

\item{by}{Type of columns to subset}

\item{biopath}{Column of info we want to compare}
}
\value{
A matrix of combinations of all the ids selected from biopath
and compare them all
}
\description{
Given a data.frame with the information and the genes we want to extract,
finds all the pathway for each gene and return a data.frame with the
combinations of all pathways of gene 1 with all pathways of gene 2
}
\note{
Documented but not exported. Used internally for \code{\link{genesSim}}
}
\examples{
library("org.Hs.eg.db")
entrezids <- keys(org.Hs.eg.db, keytype = "ENTREZID")
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- select(org.Hs.eg.db, keys = entrezids, keytype = "ENTREZID",
                     columns = "PATH")
combBiopath(c("81", "18"), genes.kegg, "ENTREZID", "PATH")
}
\author{
Lluís Revilla
}
