% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{combinadic}
\alias{combinadic}
\title{i-th combination of n elements taken from r}
\usage{
combinadic(n, r, i)
}
\arguments{
\item{n}{Elements to extract the combination from}

\item{r}{Number of elements per combination}

\item{i}{ith combination}
}
\description{
Function similar to combn but for larger vectors
}
\seealso{
\code{\link{combn}}
}

