% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar.R
\name{combinadic}
\alias{combinadic}
\title{i-th combination of n elements taken from r}
\usage{
combinadic(n, r, i)
}
\arguments{
\item{n}{Elements to extract the combination from}

\item{r}{Number of elements per combination}

\item{i}{ith combination}
}
\value{
The combination ith of the elements
}
\description{
Function similar to combn but for larger vectors. To avoid allocating a big
vector with all the combinations each one can be computed with this
function.
}
\examples{
#Output of all combinations
combn(LETTERS[1:5], 2)
# Otuput of the second combination
combinadic(LETTERS[1:5], 2, 2)
}
\references{
\href{http://stackoverflow.com/a/4494469/2886003}{StackOverflow answer
4494469/2886003}
}
\seealso{
\code{\link{combn}}
}
\author{
Joshua Ulrich
}
