% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineScores.R
\name{combineScores}
\alias{combineScores}
\title{Combining values}
\usage{
combineScores(scores, method, round = FALSE)
}
\arguments{
\item{scores}{Matrix of scores to be combined}

\item{method}{one of \code{c("avg", "max", "rcmax", "rcmax.avg", "BMA")}
see details}

\item{round}{Should the resulting value be rounded to the third digit?}
}
\value{
A numeric value as described in details.
}
\description{
Combine several values into one by several methods.
}
\details{
The methods return:
\describe{\item{avg}{The average or mean value}
\item{max}{The max value}
\item{rcmax}{The max of the column means or row means}
\item{rcmax.avg}{The sum of the max values by rows and columns divided by
the number of columns and rows}
\item{BMA}{The same as \code{rcmax.avg}}}
}
\note{
This is a version of combineScores inspired from GOSemSim
\code{\link[GOSemSim]{combineScores}} with optional rounding and some
internal differences.
}
\examples{
d <- structure(c(0.4, 0.6, 0.222222222222222, 0.4, 0.4, 0, 0.25, 0.5,
0.285714285714286), .Dim = c(3L, 3L), .Dimnames = list(c("a",
"b", "c"), c("d", "e", "f")))
d
sapply(c("avg", "max", "rcmax", "rcmax.avg", "BMA"), combineScores,
       scores = d)
d[1,2] <- NA
sapply(c("avg", "max", "rcmax", "rcmax.avg", "BMA"), combineScores,
       scores = d)
}
\author{
Lluís Revilla based on Guan
}
