% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{compare_graphs}
\alias{compare_graphs}
\title{Compare graph structure}
\usage{
compare_graphs(g1, g2)
}
\arguments{
\item{g1, g2}{Graph in GraphNEL format, or a character list}
}
\description{
Function to estimate how much two graphs overlap by looking how much of the
nodes are shared. It also works with list of genes.
}

