% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-incidence.R
\name{distribution}
\alias{distribution}
\title{Distribution of genes per pathway in pathways}
\usage{
distribution(object)
}
\arguments{
\item{object}{A GeneSetCollection.}
}
\value{
A matrix with the number of pathways per genes as columns and the
genes per pathways as rows.
}
\description{
Calculates how many pathways per gene are distributed in genes per pathways.
}
\author{
Lluís Revilla
}
