% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-methods.R
\name{gene}
\alias{gene}
\title{Statistics about genes}
\usage{
gene(object, gene)
}
\arguments{
\item{object}{A GeneSetCollection object}

\item{gene}{A character of the gene to analyze. If missing it will analyze
all genes}
}
\value{
If a single gene is provided invisible a list of statistics (the
statistics will be printed on the screen), otherwise the data.frame of
statistics.
}
\description{
Calculates some statistics of the genes, like the biggest
pathway, the gene in more pathways, h-index and entropy or information
content for the genes per pathway and pathways for genes and they percentage
of the maximum.
}
\author{
Lluís Revilla
}
