% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSim.R
\name{geneSim}
\alias{geneSim}
\alias{mgeneSim}
\title{Similarity score genes based on pathways similarity}
\usage{
geneSim(gene1, gene2, info, method = "max", ...)

mgeneSim(genes, info, method = "max", ...)
}
\arguments{
\item{gene1, gene2}{Ids of the genes to calculate the similarity, to be found
in genes.}

\item{info}{A list of genes and the pathways they are involved.}

\item{method}{To combine the scores of each pathway, one of \code{c("avg",
"max", "rcmax", "rcmax.avg", "BMA")}, if NULL returns the matrix of
similarities.}

\item{...}{Other arguments passed to \code{\link{combineScores}}}

\item{genes}{A vector of genes.}
}
\value{
The highest Dice score of all the combinations of pathways between
the two ids compared if a method to combine scores is provided or NA if
there isn't information for one gene.
If an \code{NA} is returned this means that there isn't information
available for any pathways for one of the genes. Otherwise a number
between 0 and 1 (both included) is returned. Note that there isn't a
negative value of similarity.

\code{mgeneSim} returns the matrix of similarities between the genes
in the vector
}
\description{
Given two genes, calculates the Dice similarity between each pathway
which is combined to obtain a similarity between the genes.
}
\details{
Given the information about the genes and their pathways, uses the ids
of the genes to find the Dice similarity score for each pathway comparison
between the genes. Later this similarities are combined using
\code{\link{combineScores}}.
}
\note{
genes accept named characters and the output will use the names of the
genes.
}
\examples{
library("org.Hs.eg.db")
library("reactome.db")
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- as.list(org.Hs.egPATH)
# Extracts the paths of all genes of org.Hs.eg.db from reactome
genes.react <- as.list(reactomeEXTID2PATHID)
geneSim("81", "18", genes.react)
geneSim("81", "18", genes.kegg)
geneSim("81", "18", genes.react, NULL)
geneSim("81", "18", genes.kegg, NULL)

mgeneSim(c("81", "18", "10"), genes.react)
mgeneSim(c("81", "18", "10"), genes.react, "avg")
named_genes <- structure(c("81", "18", "10"),
.Names = c("ACTN4", "ABAT", "NAT2"))
mgeneSim(named_genes, genes.react, "max")
}
\seealso{
\code{\link{conversions}} help page to transform Dice
score to Jaccard score. For the method to combine the scores see
\code{\link{combineScores}}.
}
\author{
Lluis Revilla
}
