% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\docType{methods}
\name{GenesPerPathway}
\alias{GenesPerPathway}
\alias{GenesPerPathway,GeneSetCollection-method}
\title{Genes per pathway}
\usage{
GenesPerPathway(object)

\S4method{GenesPerPathway}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
A list of genes per pathway.
}
\description{
Calculates the genes per pathway of a GeneSetCollection
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Calculates genes per pathway of the
GeneSetCollection
}}

\seealso{
\code{\link{PathwaysPerGene}}
}
\author{
Lluís Revilla
}
