% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{genesSim}
\alias{genesSim}
\alias{mgeneSim}
\title{Calculates the Dice similarity score of two genes}
\usage{
genesSim(gene1, gene2, genes, id, pathwayDB, method = "max")

mgeneSim(gene.list, genes, id, pathwayDB, method = "max")
}
\arguments{
\item{gene1, gene2}{Entrez gene id.}

\item{genes}{is the matrix with the information to calculate the similarity.
It is created using select, should contain the column "id" and "pathwayDB".}

\item{id}{is the column of \code{genes} where \code{gene1} and \code{gene2}
are to be found}

\item{pathwayDB}{is the column where pathways should be found. It is usually
the name of the database where they come from.}

\item{method}{To combine the scores of each pathway, one of \code{c("avg",
"max", "rcmax", "rcmax.avg", "BMA")}, if NULL returns the matrix.}

\item{gene.list}{Given a list of vectors return the similarities}
}
\value{
The highest Dice score of all the combinations of pathways between
the two ids compared if a method to combine scores is provided or NA if
there isn't information for one gene.
If an \code{NA} is returned this means that there isn't information
available for any pathways for one of the genes. Otherwise a number
between 0 and 1 (both included) is returned. Note that there isn't a
negative value of similarity.

\code{mgeneSim} returns the matrix of similarities between the genes
in the vector
}
\description{
For the genes given calculates the Dice similarity between each pathway
which is combined to obtain a similarity between the genes.
}
\details{
Given the information about the genes and their pathways, uses the ids
of the genes to find the Dice similarity score for each pathway comparison
between the genes. Later this similarities are combined using
\code{\link{combineScores}}.
}
\examples{
library("org.Hs.eg.db")
library("reactome.db")
entrezids <- keys(org.Hs.eg.db, keytype = "ENTREZID")
#Extract the paths of all genes of org.Hs.eg.db from KEGG (last update in
# data of June 31st 2011)
genes.kegg <- select(org.Hs.eg.db, keys = entrezids, keytype = "ENTREZID",
                     columns = "PATH")
# Extracts the paths of all genes of org.Hs.eg.db from reactome
genes.react <- select(reactome.db, keys = entrezids, keytype = "ENTREZID",
                      columns = "REACTOMEID")
genesSim("81", "18", genes.react, "ENTREZID", "REACTOMEID")
genesSim("81", "18", genes.kegg, "ENTREZID", "PATH")
genesSim("81", "18", genes.react, "ENTREZID", "REACTOMEID", NULL)

mgeneSim(c("81", "18", "10"), genes.react, "ENTREZID", "REACTOMEID")
mgeneSim(c("81", "18", "10"), genes.react, "ENTREZID", "REACTOMEID", "avg")
}
\seealso{
\code{\link{conversions}} help page to transform Dice
score to Jaccard score. For the method to combine the scores see
\code{\link{combineScores}}.
}
\author{
Lluis Revilla
}
