% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathSim.R
\docType{methods}
\name{incidence,list-method}
\alias{incidence,list-method}
\title{Creates the incidence matrix}
\usage{
\S4method{incidence}{list}(x)
}
\arguments{
\item{x}{A list}
}
\value{
A matrix with pathways as rows and genes in columns.
}
\description{
Given a list of pathways and its genes creates an incidence matrix
}
\author{
Lluís Revilla
}
