% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{inverseList}
\alias{inverseList}
\title{Invert a list}
\usage{
inverseList(x)
}
\arguments{
\item{incidence}{Incidence matrix}
}
\value{
The number of pathways each gene has
}
\description{
Calculate the pathways per gene of an incidence matrix
}
\keyword{internal}
