% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathSim.R
\name{pathSim}
\alias{pathSim}
\alias{mpathSim}
\title{Calculates the Dice similarity between pathways}
\usage{
pathSim(pathway1, pathway2, info)

mpathSim(pathways, info, method = "max", ...)
}
\arguments{
\item{pathway1, pathway2}{A single pathway to calculate the similarity}

\item{info}{A list of genes and the pathways they are involved.}

\item{pathways}{Pathways to calculate the similarity for}

\item{method}{To combine the scores of each pathway, one of \code{c("avg",
"max", "rcmax", "rcmax.avg", "BMA")}, if NULL returns the matrix of
similarities.}

\item{...}{Other arguments passed to \code{\link{combineScores}}}
}
\value{
The similarity between those pathways or all the similarities
between each comparison.
}
\description{
Calculates the similarity between pathways using dice similarity score.
}
\details{
\code{diceSim} is used to calculate similarities between the two pathways.

\code{mpathSim} compares the similarity between several pathways and can use
\code{\link{combineScores}} to extract the similarity between those
pathways.
If one needs the matrix of similarities between pathways set the argument
methods to \code{NULL}.
}
\examples{
library("reactome.db")
# Extracts the paths of all genes of org.Hs.eg.db from reactome
genes.react <- as.list(reactomeEXTID2PATHID)
pathways <- c("112315", "112310", "112316", "373753", "916853", "109582",
"114608", "1500931")
pathSim("112310", "112316", genes.react)
mpathSim(pathways, genes.react, NULL)
}
\seealso{
\code{\link{diceSim}} and \code{\link{combineScores}} and
\code{\link{conversions}} help page to transform Dice score to Jaccard
score.
}
\author{
Lluís Revilla
}
