% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-methods.R
\docType{methods}
\name{PathwaysPerGene}
\alias{PathwaysPerGene}
\alias{PathwaysPerGene,GeneSetCollection-method}
\title{Pathways per gene}
\usage{
PathwaysPerGene(object)

\S4method{PathwaysPerGene}{GeneSetCollection}(object)
}
\arguments{
\item{object}{A GeneSetCollection object}
}
\value{
a list of pathways per genes
}
\description{
Calculates the pathways per gene of a GeneSetCollection
}
\section{Methods (by class)}{
\itemize{
\item \code{GeneSetCollection}: Pathways per gene in the GeneSetCollection
}}

\seealso{
\code{\link{GenesPerPathway}}
}
\author{
Lluís Revilla
}
