% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{seq2mat}
\alias{seq2mat}
\title{Transform a vector to a symetric matrix}
\usage{
seq2mat(x, dat)
}
\arguments{
\item{x}{names of columns and rows, used to define the sieze of the matrix}

\item{dat}{Data to fill with the matrix}
}
\description{
The matrix should be of \code{ncol = length(x)} and \code{nrow = length(x)},
so \code{dat} is at least \code{choose(length(x), 2)} of length.
}
\details{
It assumes that the data provided comes from using the row and column id to
obtain it.
}
\examples{
seq2mat(LETTERS[1:5], 1:10)
}

