% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{similarities}
\alias{similarities}
\title{Similarities}
\usage{
similarities(sim, func, ...)
}
\arguments{
\item{sim}{list of similarities to be joined}

\item{func}{function to perform on those similarities}

\item{...}{Other arguments passed to the function \code{func}.}
}
\value{
A matrix of the size of the similarities
}
\description{
Function to join list of similarities by a function provided by the user.
}

