% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cor.R
\name{weighted}
\alias{weighted}
\title{Calculates the weighted sum of the values}
\usage{
weighted(x, weights)
}
\arguments{
\item{x}{Vector of numbers}

\item{weights}{Vector of weights}
}
\value{
A number product of x*weights removing all \code{NA} values
}
\description{
Each values should have its weight, otherwise it will throw an error.
}

